+++ src/lib/items/Item.js
@@ -32,7 +32,8 @@ export default class Item extends Component {
     // onDrop: React.PropTypes.func,
     // onResizing: React.PropTypes.func,
     // onResized: React.PropTypes.func,
-    // onContextMenu: React.PropTypes.func
+    // onContextMenu: React.PropTypes.func,
+    // contentComponent: React.PropTypes.func
   }
   static defaultProps = {
     selected: false
@@ -421,6 +422,15 @@ export default class Item extends Component {
     }
   }
 
+  renderContent() {
+    const ContentComponent = this.props.contentComponent
+    if(ContentComponent) {
+      return <ContentComponent item={this.props.item} />
+    } else {
+      return this.itemTitle
+    }
+  }
+
   render () {
     const dimensions = this.props.dimensions
     if (typeof this.props.order === 'undefined' || this.props.order === null) {
@@ -461,7 +471,7 @@ export default class Item extends Component {
         {this.props.useResizeHandle ? <div ref='dragLeft' className='rct-drag-left'></div> : ''}
         <div className='rct-item-overflow'>
           <div className='rct-item-content'>
-            {this.itemTitle}
+            {this.renderContent()}
           </div>
         </div>
         {this.props.useResizeHandle ? <div ref='dragRight' className='rct-drag-right'></div> : ''}